Add files via upload
Coding Academy: Sprint 1.
Base version of the Minesweeper game.
Not very satisfied with the structure of the code. I feel it could be a lot cleaner and a lot more logical. But at least the game works for what it's supposed to do.
Before implementing the bonuses I am considering starting over from scratch to over a more solid base game foundation.